Move distribute tests out of contrib dir.
PiperOrigin-RevId: 241643770
