+++ tensorflow/contrib/distribute/python/monitor_test.py
@@ -21,11 +21,11 @@ from __future__ import print_function
 from absl.testing import parameterized
 import numpy
 from tensorflow.contrib.distribute.python import monitor as monitor_lib
-from tensorflow.contrib.distribute.python.single_loss_example import single_loss_example
 from tensorflow.python.client import session
 from tensorflow.python.distribute import combinations
 from tensorflow.python.distribute import one_device_strategy
 from tensorflow.python.distribute import strategy_combinations
+from tensorflow.python.distribute.single_loss_example import single_loss_example
 from tensorflow.python.eager import context
 from tensorflow.python.eager import test
 from tensorflow.python.framework import ops
