+++ tensorflow/python/distribute/moving_averages_test.py
