+++ tensorflow/python/distribute/step_fn_test.py
@@ -20,9 +20,9 @@ from __future__ import print_function
 
 from absl.testing import parameterized
 import numpy
-from tensorflow.contrib.distribute.python.single_loss_example import single_loss_example
 from tensorflow.python.distribute import combinations
 from tensorflow.python.distribute import strategy_combinations
+from tensorflow.python.distribute.single_loss_example import single_loss_example
 from tensorflow.python.eager import context
 from tensorflow.python.eager import test
 from tensorflow.python.ops import variables
