+++ tensorflow/python/distribute/strategy_test_lib.py
