+++ tensorflow/python/keras/distribute/keras_embedding_model_correctness_test.py
@@ -18,10 +18,10 @@ from __future__ import division
 from __future__ import print_function
 
 import numpy as np
-from tensorflow.contrib.distribute.python import keras_correctness_test_base
 from tensorflow.python import keras
 from tensorflow.python.distribute import combinations
 from tensorflow.python.eager import test
+from tensorflow.python.keras.distribute import keras_correctness_test_base
 from tensorflow.python.training import gradient_descent
