+++ tensorflow/python/keras/distribute/keras_utils_test.py
@@ -22,11 +22,11 @@ import collections
 import tempfile
 from absl.testing import parameterized
 import numpy as np
-from tensorflow.contrib.distribute.python import tpu_strategy
 from tensorflow.python import keras
 from tensorflow.python.data.ops import dataset_ops
 from tensorflow.python.distribute import combinations
 from tensorflow.python.distribute import strategy_combinations
+from tensorflow.python.distribute import tpu_strategy
 from tensorflow.python.distribute import values
 from tensorflow.python.eager import test
 from tensorflow.python.framework import constant_op
